requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
colorama==0.4.4
    # via rich
commonmark==0.9.1
    # via rich
html2text==2020.1.16
    # via -r requirements.in
mysql-connector-python==8.0.25
    # via -r requirements.in
protobuf==3.17.2
    # via mysql-connector-python
pydantic==1.8.2
    # via -r requirements.in
pygments==2.9.0
    # via rich
python-slugify==5.0.2
    # via -r requirements.in
rich==10.2.2
    # via -r requirements.in
six==1.16.0
    # via protobuf
text-unidecode==1.3
    # via python-slugify
typing-extensions==3.10.0.0
    # via
    #   pydantic
    #   rich
